third_party/WebKit/LayoutTests/http/tests/w3c/webperf/submission/Intel/resource-timing/test_resource_timing_timing_allow_cross_origin_resource_request-expected.txt

CONSOLE MESSAGE: line 24: Uncaught TypeError: Cannot read property 'redirectStart' of undefined
Description

This test validates the values in resource timing for a timing-allow-origin cross-origin resource request.




Harness Error. harness_status.status = 2 , harness_status.message = null
